src/requirements.txt
scrapy==2.8.0
scrapy-autounit==0.0.32
dateparser==1.1.7
mysql_connector_python==8.0.32